% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dat.R
\name{abif_dat}
\alias{abif_dat}
\title{Creates an \code{abif_dat} object, which represents a data element from an ABIF
file.}
\usage{
abif_dat(data, type)
}
\arguments{
\item{data}{atomic vector containing the data}

\item{type}{integer. The element type of the data}
}
\value{
abif_dat object.
}
\description{
Creates an \code{abif_dat} object, which represents a data element from an ABIF
file.
}
