% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{abif_dir}
\alias{abif_dir}
\title{Create a new ABIF directory object from raw data.}
\usage{
abif_dir(dir_offset, file_raw, keep_data = c("parsed", "raw", "both", "none"))
}
\arguments{
\item{dir_offset}{integer. Offset of the directory entry in file_raw.}

\item{file_raw}{raw. Data containing the directory entry.}

\item{keep_data}{character. Specifies what directory entry data to keep
(raw, parsed, both, or none).}
}
\value{
abif_dir object containing the directory entry information.
}
\description{
Create a new ABIF directory object from raw data.
}
