% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_head.R
\name{abif_head}
\alias{abif_head}
\title{Creates a new \code{abif_head} object, which is used to represent the header of
an ABIF file, such as .fsa or .hid.}
\usage{
abif_head(dat)
}
\arguments{
\item{dat}{raw. The data containing the ABIF file header. Must be at least
the first 30 bytes of the file.}
}
\value{
abif_head object containing the parsed header data.
}
\description{
Creates a new \code{abif_head} object, which is used to represent the header of
an ABIF file, such as .fsa or .hid.
}
