% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_peak.R
\name{abif_peaks}
\alias{abif_peaks}
\title{Creates a data.frame from the \code{Peak} data in an ABIF file.}
\usage{
abif_peaks(x, friendly_names = TRUE, dye_names = TRUE)
}
\arguments{
\item{x}{hid or abif_dir object.}

\item{friendly_names}{logical. Whether to rename columns with more
meaningful names or use the directory entry names.}

\item{dye_names}{logical. Whether to add dye names to the output data.frame.}
}
\value{
data.frame created from the \code{Peak} entries in x.
}
\description{
Creates a data.frame from the \code{Peak} data in an ABIF file.
}
