% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{char}
\alias{char}
\title{Parses the char element type of ABIF files from raw data.}
\usage{
char(f)
}
\arguments{
\item{f}{Raw vector}
}
\value{
Character vector
}
\description{
Parses the char element type of ABIF files from raw data.
}
