% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_element_size}
\alias{dir_element_size}
\title{Extracts the element size of an ABIF directory object.}
\usage{
dir_element_size(directory)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}
}
\value{
integer. The directory entry element size in bytes.
}
\description{
Extracts the element size of an ABIF directory object.
}
