% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_n}
\alias{dir_n}
\title{Extracts the number of elements from an ABIF directory object.}
\usage{
dir_n(directory)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}
}
\value{
integer. The directory entry's number of elements.
}
\description{
Extracts the number of elements from an ABIF directory object.
}
