% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_name}
\alias{dir_name}
\title{Extracts the name from an ABIF directory object.}
\usage{
dir_name(directory, include_num = FALSE)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}

\item{include_num}{logical. Whether to include the directory entry number.}
}
\value{
character. The directory entry name.
}
\description{
When include_num == TRUE, includes the directory entry number in the output
in the following format: "Name.Num". When FALSE (the default), only the name
is returned.
}
