% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_num}
\alias{dir_num}
\title{Extracts the number of an ABIF directory object.}
\usage{
dir_num(directory)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}
}
\value{
integer. The directory entry number.
}
\description{
Extracts the number of an ABIF directory object.
}
