% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_offset}
\alias{dir_offset}
\title{Extracts the file offset of an ABIF directory object.}
\usage{
dir_offset(directory)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}
}
\value{
integer. The directory entry's file offset.
}
\description{
Extracts the file offset of an ABIF directory object.
}
