% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_size}
\alias{dir_size}
\title{Extracts the data size of an ABIF directory object.}
\usage{
dir_size(directory)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}
}
\value{
integer. The directory entry's data size in bytes.
}
\description{
Extracts the data size of an ABIF directory object.
}
