% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_type}
\alias{dir_type}
\title{Extracts the element type of an ABIF directory object.}
\usage{
dir_type(directory)
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}
}
\value{
integer. The directory entry element type.
}
\description{
Extracts the element type of an ABIF directory object.
}
\seealso{
\code{\link[=element_types]{element_types()}}
}
