% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_peak.R
\name{dye_names}
\alias{dye_names}
\title{Extracts a list of the dye names and their corresponding index values from an
\code{abif_dir} object.}
\usage{
dye_names(directory)
}
\arguments{
\item{directory}{abif_dir object}
}
\value{
list with elements \code{dye_index} and \code{dye_name}.
}
\description{
Extracts a list of the dye names and their corresponding index values from an
\code{abif_dir} object.
}
