% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{extract_directory_entry}
\alias{extract_directory_entry}
\title{Extracts the information for a single ABIF directory entry. This function returns
correct values for standard directory entries but will return incorrect information
for special data types.}
\usage{
extract_directory_entry(raw_data)
}
\arguments{
\item{raw_data}{Raw vector of length 28 representing the directory entry.}
}
\value{
List with elements corresponding to the ABIF directory entry information.
}
\description{
Extracts the information for a single ABIF directory entry. This function returns
correct values for standard directory entries but will return incorrect information
for special data types.
}
