% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{extract_header}
\alias{extract_header}
\title{Extracts the ABIF header}
\usage{
extract_header(raw_data)
}
\arguments{
\item{raw_data}{Raw vector with contents of .fsa or .hid file.}
}
\value{
List with elements corresponding to the ABIF header information.
}
\description{
Extracts the ABIF header
}
