% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid_data.R
\name{extract_raw_data}
\alias{extract_raw_data}
\title{Extracts the raw data associated with an ABIF directory entry.}
\usage{
extract_raw_data(raw_data, data_offset, data_size)
}
\arguments{
\item{raw_data}{Raw vector representing the .fsa or .hid file.}

\item{data_offset}{Integer vector of length 1. The file offset for the data of the directory entry.}

\item{data_size}{Integer vector of length 1. The data size of the directory entry.}
}
\value{
Raw vector containing the data for the directory entry.
}
\description{
Extracts the raw data associated with an ABIF directory entry.
}
