% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{float32}
\alias{float32}
\title{Parses 32 bit floating numbers from a raw vector}
\usage{
float32(f, n = 1L, ...)
}
\arguments{
\item{f}{Raw vector}

\item{n}{The maximum number of records to be read. Passed to \code{readBin()}.}

\item{...}{Other arguments to pass to \code{readBin()}}
}
\value{
Double vector
}
\description{
Parses 32 bit floating numbers from a raw vector
}
