% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{hid_data}
\alias{hid_data}
\title{Extracts the \code{data} element from an hid object. If \code{pattern} is specified, returns only the elements of \code{data} that match pattern using \code{grep()}.}
\usage{
hid_data(x, pattern = NULL, what = c("parsed", "raw", "both"), ...)
}
\arguments{
\item{x}{hid object}

\item{pattern}{character. A string containing the data name to search for. Passed to \code{grep()}.}

\item{what}{character. Specifies what directory entry data to keep.}

\item{...}{Other arguments passed to \code{grep()}}
}
\value{
list. Invisibly returns the data element of an hid object.
}
\description{
Extracts the \code{data} element from an hid object. If \code{pattern} is specified, returns only the elements of \code{data} that match pattern using \code{grep()}.
}
\examples{
\dontrun{
my_hid_data <- hid_data(my_hid)
# Get the name of the first dye
my_hid_data$DyeN.1
# Get the instrument model name
my_hid_data$MODL.1
# Get all the data elements with "DATA" in the name
hid_data(my_hid, "DATA")
}
}
