% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{hid_peaks}
\alias{hid_peaks}
\title{Extracts the \code{hid_peaks} element from an hid object.}
\usage{
hid_peaks(x)
}
\arguments{
\item{x}{hid object}
}
\value{
dataframe. Invisibly returns the \code{hid_peaks} element from the hid object
}
\description{
Extracts the \code{hid_peaks} element from an hid object.
}
\examples{
\dontrun{
my_hid_peaks <- hid_peaks(my_hid)
# Calculate average peak height
mean(my_hid_peaks$height)
}
}
