% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{hidTime}
\alias{hidTime}
\title{Parses time field from raw data.}
\usage{
hidTime(f)
}
\arguments{
\item{f}{Raw vector}
}
\value{
List of the hour, minute, second, and hsecond data
}
\description{
Parses time field from raw data.
}
