% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{int8}
\alias{int8}
\title{Parse signed 8 bit integers from a raw vector}
\usage{
int8(f, n = 1L, ...)
}
\arguments{
\item{f}{Raw vector#'}

\item{n}{The maximum number of records to be read. Passed to \code{readBin()}.}

\item{...}{Other arguments to pass to \code{readBin()}}
}
\value{
Integer vector
}
\description{
Parse signed 8 bit integers from a raw vector
}
