% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dat.R
\name{parse_dat}
\alias{parse_dat}
\title{Parses the raw data associated with an ABIF data object (\code{abif_dat}).}
\usage{
parse_dat(x)
}
\arguments{
\item{x}{abif_dat object containing raw data.}
}
\value{
list, character, or numeric vector, depending on element type.
The parsed data from the directory entry.
}
\description{
Parses the raw data associated with an ABIF data object (\code{abif_dat}).
}
