% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid_data.R
\name{parse_data}
\alias{parse_data}
\title{Parses the raw data associated with an ABIF directory entry.}
\usage{
parse_data(raw_data, element_type, num_elements)
}
\arguments{
\item{raw_data}{Raw vector of data from the directory entry.}

\item{element_type}{Integer vector of length 1. The element type of the directory entry.}

\item{num_elements}{Integer vector of length 1. The number of elements of the directory entry.}
}
\value{
List, character, or numeric vector, depending on element type. The parsed data from the directory entry.
}
\description{
Parses the raw data associated with an ABIF directory entry.
}
