% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{pString}
\alias{pString}
\title{Parses the pString element type of ABIF files from raw data.}
\usage{
pString(f)
}
\arguments{
\item{f}{Raw vector}
}
\value{
Character vector
}
\description{
Parses the pString element type of ABIF files from raw data.
}
