% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{read_hid}
\alias{read_hid}
\title{Reads an hid file}
\usage{
read_hid(filepath, ...)
}
\arguments{
\item{filepath}{Character vector of length 1 containing path to the .fsa or .hid file.}

\item{...}{Additional arguments to pass to \code{readBin()}}
}
\value{
Raw vector with contents of the file.
}
\description{
Reads an hid file
}
