% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readhid-package.R
\docType{package}
\name{readhid-package}
\alias{readhid}
\alias{readhid-package}
\title{readhid: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SAR85/readhid}
  \item Report bugs at \url{https://github.com/SAR85/readhid/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
