% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{rtc}
\alias{rtc}
\title{Parses strings from raw vector}
\usage{
rtc(x, ...)
}
\arguments{
\item{x}{raw vector}

\item{...}{Other arguments passed to \code{rawToChar()}}
}
\value{
Character vector
}
\description{
Parses strings from raw vector
}
