% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid_data.R
\name{special_types}
\alias{special_types}
\title{Returns vector of the "special" element types found in .hid files}
\usage{
special_types()
}
\value{
Vector of integers
}
\description{
Returns vector of the "special" element types found in .hid files
}
