% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{uint32}
\alias{uint32}
\title{Parses unsigned 32 bit integers from a raw vector}
\usage{
uint32(f, n = 1L, ...)
}
\arguments{
\item{f}{Raw vector}

\item{n}{The maximum number of records to be read. Passed to \code{readBin()}.}

\item{...}{Other arguments to pass to \code{readBin()}}
}
\value{
Integer vector
}
\description{
Parses unsigned 32 bit integers from a raw vector
}
