% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_data}
\alias{dir_data}
\title{#' Extracts data from ABIF directory entries. This is a generic function so
methods can support objects such as hid, list, and individual directory
entries (abif_dir).}
\usage{
dir_data(x, pattern = NULL, what = c("parsed", "raw"), ...)
}
\arguments{
\item{x}{object from which to extract data.}

\item{pattern}{character. A string containing the directory entry name(s)
to search for. Passed to \code{grep()}.}

\item{what}{character. Specifies what directory entry data to keep ("parsed"
or "raw").}

\item{...}{Other arguments passed to \code{grep()}}
}
\value{
invisibly returns the matching data element(s). Object type matches
the form of \code{x}: an hid object or list will return a list and an abif_dir
object will return a vector.
}
\description{
If \code{pattern} is specified, returns data from only the directory entries with
names that match \code{pattern} using \code{grep()}.
}
\examples{
\dontrun{
# Get the name of the first dye from an hid object
dir_data(my_hid, pattern = "DyeN.1", what = "parsed")
# Get all the data elements with "DATA" in the name
dir_data(my_hid, pattern = "DATA")
}
}
