% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dir.R
\name{dir_data}
\alias{dir_data}
\title{Extracts the data of an ABIF directory object.}
\usage{
dir_data(directory, what = c("parsed", "raw", "both"))
}
\arguments{
\item{directory}{abif_dir. The ABIF directory object.}

\item{what}{character. Specifies what directory entry data to keep.
(raw, parsed, both).}
}
\value{
list containing the directory entry's data
}
\description{
Extracts the data of an ABIF directory object.
}
\seealso{
\code{\link[=parse_data]{parse_data()}}
}
