% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dat.R
\name{element_types}
\alias{element_types}
\title{Returns the name, element size, substitute element type, or the name of the
function used to parse the element type.}
\usage{
element_types(element_type, get = c("name", "size", "sub", "parse_fun"))
}
\arguments{
\item{element_type}{integer representing the element type to look up.}

\item{get}{character specifying the data to return.}
}
\value{
character or integer, depending on value of \code{get}
}
\description{
Return value depends on the \code{get} parameter:
"name" returns a string with the name of the element type
"size" returns an integer with the element size for the element type\
"sub" returns an integer with the standard element type substitute. For
special element types, this will be the standard element type used to
parse the data. For normal element types, this will be the same as the
element type.
"parse_fun" returns a string with the name of the function used to parse the
element type.
}
\examples{
# Element type 33 is a special type. The equivalent standard type can be found
# using "sub":
element_types(33L, "sub")

# Retrieve the parsing function for an element type:
parse_fun <- match.fun(element_types(13L, "parse_fun"))
identical(parse_fun, as.logical)
parse_fun(as.raw(0x01), n = 1)

}
\seealso{
\code{\link[=special_types]{special_types()}}
}
