% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid_data.R
\name{element_types}
\alias{element_types}
\title{Returns the name, element type, element size, or substitute element type for
a given element type.}
\usage{
element_types(element_type, get = c("name", "type", "size", "sub"))
}
\arguments{
\item{element_type}{Integer vector of length 1 to look up}

\item{get}{Character vector of length 1 specifying the data to return}
}
\value{
Character vector or integer vector of length 1. Depends on value of \code{get}
}
\description{
Returns the name, element type, element size, or substitute element type for
a given element type.
}
