% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{extract_directory}
\alias{extract_directory}
\title{Extracts the ABIF directory.}
\usage{
extract_directory(raw_data, directory_offset, num_elements)
}
\arguments{
\item{raw_data}{Raw vector containing the contents of a .fsa or .hid file.}

\item{directory_offset}{The 0-based offset of the directory in the file. This is not the offset in the raw_data vector, which would larger by 1 due to R's 1-based indexing.}

\item{num_elements}{Integer containing the number of directory entries.}
}
\value{
List containing a list for each directory entry.
}
\description{
Extracts the ABIF directory.
}
