% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{hid}
\alias{hid}
\title{Creates an hid object with the data from an .fsa or .hid file.}
\usage{
hid(filepath, raw = FALSE)
}
\arguments{
\item{filepath}{character vector of length 1. Path to the .fsa or .hid file.}

\item{raw}{logical. When TRUE, returns the raw data for each directory entry instead of the parsed data.}
}
\value{
hid object
}
\description{
Creates an hid object with the data from an .fsa or .hid file.
}
\examples{
\dontrun{
my_hid_file <- hid("/path/to/file.hid")
}

}
