% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{hid}
\alias{hid}
\title{Creates an \code{hid} object}
\usage{
hid(filepath, keep_data = c("parsed", "raw", "both", "none"))
}
\arguments{
\item{filepath}{character. Path to the .fsa or .hid file.}

\item{keep_data}{character. Specifies what directory entry data to keep
(raw, parsed, both, or none). If "raw" or "both", also keeps the entire file
raw data.}
}
\value{
hid object
}
\description{
\code{hid} objects represent the data in an Applied Biosystems Inc. Format (ABIF)
file. These files contain DNA sequencing or fragment analysis data produced
by Genetic Analyzers and commonly have file extensions such as .fsa or .hid.
}
\examples{
\dontrun{
my_hid_file <- hid("/path/to/file.hid", keep_data = "parsed")
}

}
