% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{hidDate}
\alias{hidDate}
\title{Parses a date from a raw vector}
\usage{
hidDate(f, ...)
}
\arguments{
\item{f}{Raw vector}

\item{...}{Other arguments passed to \code{as.Date()}}
}
\value{
Date vector of length 1
}
\description{
Parses a date from a raw vector
}
