% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dat.R
\name{parse_dat}
\alias{parse_dat}
\title{Parses the raw data associated with an ABIF data object (\code{abif_dat}).}
\usage{
parse_dat(x)
}
\arguments{
\item{x}{abif_dat object containing raw data.}
}
\value{
atomic vector, object, or list, containing the parsed data from the
directory entry.
}
\description{
Times (element type 11) are returned as a list and dates (element
type 10) are returned as a \code{Date}. Other element types are returned as
numeric or character vectors. If x does not contain raw data, the data is
returned unchanged.
}
\examples{
my_dat <- abif_dat(as.raw(0x01), type = 13L)
x <- parse_dat(my_dat)

}
