% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid.R
\name{read_hid}
\alias{read_hid}
\title{Read an ABIF file}
\usage{
read_hid(filepath, ...)
}
\arguments{
\item{filepath}{character. Path to the ABIF file.}

\item{...}{Additional arguments passed to \code{readBin()}}
}
\value{
raw. The raw contents of the file.
}
\description{
This function reads the raw data from an ABIF file.
}
