% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hid_data.R
\name{seek_raw}
\alias{seek_raw}
\title{Searches a raw vector containing a .fsa or .hid file for a directory entry
name and number, returning the index where found.}
\usage{
seek_raw(name, number, raw_data)
}
\arguments{
\item{name}{Character vector of length 1 with the directory name.}

\item{number}{Integer vector of length 1 with the directory number.}

\item{raw_data}{Raw vector containing the data from a .fsa or .hid file}
}
\value{
integer representing the matching index in \code{raw_data}
}
\description{
Searches a raw vector containing a .fsa or .hid file for a directory entry
name and number, returning the index where found.
}
