% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{special_array_length}
\alias{special_array_length}
\title{Extracts the array length for special element types.}
\usage{
special_array_length(offset, raw_data)
}
\arguments{
\item{offset}{Integer vector of length 1 representing the file offset for the #' start of the directory entry of interest. This can be found using \code{seek_raw()}.}

\item{raw_data}{Raw vector containing data from a .fsa or .hid file.}
}
\value{
Integer vector of length 1 representing the array length.
}
\description{
Extracts the array length for special element types.
}
