% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_helpers.R
\name{special_data_offset}
\alias{special_data_offset}
\title{Extracts the data offset for special element types.}
\usage{
special_data_offset(raw_data, entry_name, entry_num)
}
\arguments{
\item{raw_data}{Raw vector containing data from a .fsa or .hid file.}

\item{entry_name}{Character vector of length 1. The name of the directory entry.}

\item{entry_num}{Integer vector of length 1. The number of the directory entry.}
}
\value{
Integer vector of length one. The file offset of the data for the directory entry
}
\description{
Extracts the data offset for special element types.
}
