% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abif_dat.R
\name{special_types}
\alias{special_types}
\title{Returns vector of the "special" element types found in newer version of
ABIF files, such as .hid files produced by the 3500 Genetic Analyzer.}
\usage{
special_types()
}
\value{
integer vector representing the special element types.
}
\description{
Returns vector of the "special" element types found in newer version of
ABIF files, such as .hid files produced by the 3500 Genetic Analyzer.
}
\examples{
ifelse(32L \%in\% special_types(), print("32 is special"), print("32 isn't special"))

}
